Create additional files
