+++ Center windows winui/Interop/Delegates.cs
@@ -1,0 +1,4 @@
+﻿namespace CenterWindow.Interop;
+internal static partial class NativeMethods
+{
+}
