+++ Center windows winui/Interop/Enums.cs
@@ -1,0 +1,4 @@
+﻿namespace CenterWindow.Interop;
+internal static partial class NativeMethods
+{
+}
